requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
astral==3.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
dotenv==0.9.9
Events==0.5
grpcio==1.78.0
idna==3.11
minio==7.2.20
numpy==2.2.6
opensearch-protobufs==0.19.0
opensearch-py==3.1.0
protobuf
pycparser==3.0
pycryptodome==3.23.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz
requests==2.32.5
six==1.17.0
typing_extensions==4.15.0
urllib3==2.6.3
uuid6==2025.0.1
Pillow>=10.0.0